requirements.txt
paho-mqtt==2.0.0
PySide6==6.7.0
python-dotenv==1.0.1
